Fixed several issues.

+++ src/main/java/model/Repetition.java
@@ -5,11 +5,15 @@ import support.Database;
 import exception.ModelException;
 import java.sql.*;
 import java.util.ArrayList;
+import java.util.Date;
 
 public class Repetition extends Model {
-    public static ArrayList<Repetition> getAll() throws ModelException {
+    public static ArrayList<Repetition> getAll(boolean includeDeleted) throws ModelException {
         try{
-            String query = "SELECT * FROM repetitions LEFT JOIN courses ON repetitions.course_id = courses.id LEFT JOIN teachers ON repetitions.teacher_id = teachers.id;";
+            String query = "SELECT * FROM repetitions LEFT JOIN courses ON repetitions.course_id = courses.id LEFT JOIN teachers ON repetitions.teacher_id = teachers.id";
+            if ( !includeDeleted ){
+                query += " WHERE repetitions.deleted_at IS NULL";
+            }
             ArrayList<Repetition> elements = new ArrayList<>();
             Connection connection = Database.getConnection();
             PreparedStatement statement = connection.prepareStatement(query);
@@ -29,7 +33,8 @@ public class Repetition extends Model {
     public static Repetition find(int id) throws ModelException {
         try{
             Repetition repetition = null;
-            String query = "SELECT * FROM repetitions LEFT JOIN courses on repetitions.course_id = courses.id LEFT JOIN teachers on repetitions.teacher_id = teachers.id WHERE repetitions.id = ?;";
+            String query = "SELECT * FROM repetitions LEFT JOIN courses on repetitions.course_id = courses.id LEFT JOIN teachers on repetitions.teacher_id = teachers.id";
+            query += " WHERE repetitions.id = ? AND repetitions.deleted_at IS NULL;";
             Connection connection = Database.getConnection();
             PreparedStatement statement = connection.prepareStatement(query);
             statement.setInt(1, id);
@@ -45,28 +50,10 @@ public class Repetition extends Model {
         }
     }
 
-    public static ArrayList<Repetition> getAllOrderedByCourse() throws ModelException {
-        try{
-            ArrayList<Repetition> elements = new ArrayList<>();
-            String query = "SELECT * FROM repetitions LEFT JOIN courses on repetitions.course_id = courses.id LEFT JOIN teachers on repetitions.teacher_id = teachers.id ORDER BY course_id;";
-            Connection connection = Database.getConnection();
-            PreparedStatement statement = connection.prepareStatement(query);
-            ResultSet resultSet = statement.executeQuery();
-            while( resultSet.next() ){
-                Repetition repetition = new Repetition();
-                repetition.setPropertiesFromResultSet(resultSet);
-                elements.add(repetition);
-            }
-            statement.close();
-            return elements;
-        }catch(SQLException ex){
-            throw new ModelException("SQL exception.", ex);
-        }
-    }
-
-    private int id = 0;
-    private Teacher teacher = null;
-    private Course course = null;
+    private int id;
+    private Course course;
+    private Date deletedAt;
+    private Teacher teacher;
 
     public int getID(){
         return this.id;
@@ -90,6 +77,10 @@ public class Repetition extends Model {
         return this;
     }
 
+    public Date getDeletedAt(){
+        return this.deletedAt;
+    }
+
     public Repetition setPropertiesFromResultSet(ResultSet resultSet) throws ModelException {
         try{
             this.id = Repetition.containsField(resultSet, "repetition_id") ? resultSet.getInt("repetition_id") : resultSet.getInt("id");
@@ -101,6 +92,7 @@ public class Repetition extends Model {
             if ( resultSet.getInt("course_id") > 0 ){
                 this.course = new Course().setPropertiesFromResultSet(resultSet);
             }
+            this.deletedAt = resultSet.getTimestamp("deleted_at");
             return this;
         }catch(SQLException ex){
             throw new ModelException("SQL exception.", ex);
@@ -135,15 +127,29 @@ public class Repetition extends Model {
         }
     }
 
-    public Repetition delete() throws ModelException {
+    public Repetition delete(boolean softDelete) throws ModelException {
         try{
             if ( this.id != 0 ){
                 Connection connection = Database.getConnection();
-                PreparedStatement statement = connection.prepareStatement("DELETE FROM repetitions WHERE id = ?;");
-                statement.setInt(1, this.id);
-                statement.executeUpdate();
+                if ( softDelete ){
+                    String query = "UPDATE lessons SET deleted_at = NOW() WHERE teacher_id = ? AND course_id = ? AND completed = FALSE;";
+                    PreparedStatement statement = connection.prepareStatement(query);
+                    statement.setInt(1, this.teacher.getID());
+                    statement.setInt(2, this.course.getID());
+                    statement.executeUpdate();
+                    statement.close();
+                    query = "UPDATE repetitions SET deleted_at = NOW() WHERE id = ?;";
+                    statement = connection.prepareStatement(query);
+                    statement.setInt(1, this.id);
+                    statement.executeUpdate();
+                    statement.close();
+                }else{
+                    PreparedStatement statement = connection.prepareStatement("DELETE FROM repetitions WHERE id = ?;");
+                    statement.setInt(1, this.id);
+                    statement.executeUpdate();
+                    statement.close();
+                }
                 this.id = 0;
-                statement.close();
             }
             return this;
         }catch(SQLException ex){
